requirements.txt
browser-use
google-generativeai
python-dotenv
playwright
langchain-google-genai
langchain-core 